Renamed models into schemas
