+++ {{cookiecutter.project_slug}}/backend/app/app/api/api_v1/endpoints/login.py
@@ -11,9 +11,9 @@ from app.core import config
 from app.core.jwt import create_access_token
 from app.core.security import get_password_hash
 from app.db_models.user import User as DBUser
-from app.models.msg import Msg
-from app.models.token import Token
-from app.models.user import User
+from app.schemas.msg import Msg
+from app.schemas.token import Token
+from app.schemas.user import User
 from app.utils import (
     generate_password_reset_token,
     send_reset_password_email,
