+++ {{cookiecutter.project_slug}}/backend/app/app/crud/user.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 from app.core.security import get_password_hash, verify_password
 from app.db_models.user import User
-from app.models.user import UserCreate, UserUpdate
+from app.schemas.user import UserCreate, UserUpdate
 
 
 def get(db_session: Session, *, user_id: int) -> Optional[User]:
