+++ {{cookiecutter.project_slug}}/backend/app/app/schemas/item.py
