+++ {{cookiecutter.project_slug}}/backend/app/app/schemas/sub_item.py
