+++ {{cookiecutter.project_slug}}/backend/app/app/schemas/user.py
