+++ {{cookiecutter.project_slug}}/backend/app/app/tests/api/api_v1/test_users.py
@@ -3,7 +3,7 @@ import requests
 from app import crud
 from app.core import config
 from app.db.session import db_session
-from app.models.user import UserCreate
+from app.schemas.user import UserCreate
 from app.tests.utils.user import user_authentication_headers
 from app.tests.utils.utils import get_server_api, random_lower_string
