+++ {{cookiecutter.project_slug}}/backend/app/app/tests/utils/item.py
@@ -1,6 +1,6 @@
 from app import crud
 from app.db.session import db_session
-from app.models.item import ItemCreate
+from app.schemas.item import ItemCreate
 from app.tests.utils.user import create_random_user
 from app.tests.utils.utils import random_lower_string
