+++ {{cookiecutter.project_slug}}/backend/app/app/tests/utils/user.py
@@ -3,7 +3,7 @@ import requests
 from app import crud
 from app.core import config
 from app.db.session import db_session
-from app.models.user import UserCreate
+from app.schemas.user import UserCreate
 from app.tests.utils.utils import random_lower_string
